+++ EmanuelCaprariu_lab2/Domain/Models/OrdersCart.cs
@@ -60,14 +60,14 @@ namespace Domain.Models
         }
         public record PlacedOrder : IOrdersCart
         {
-            internal PlacedOrder(IReadOnlyCollection<CheckOrder> checkOrder,decimal numberOfOrder, DateTime placedDate)
+            internal PlacedOrder(IReadOnlyCollection<CheckOrder> checkOrder,int numberOfOrder, DateTime placedDate)
             {
                 CheckOrder = checkOrder;
                 NumberOfOrder = numberOfOrder;
                 PlacedDate = placedDate;
             }
             public IReadOnlyCollection<CheckOrder> CheckOrder { get; }
-            public decimal NumberOfOrder { get; }
+            public int NumberOfOrder { get; }
             public DateTime PlacedDate { get; }
 
         }
